refactor: move view test utils on parent folder
